+++ CHANGELOG.md
@@ -24,6 +24,7 @@ It is recommend to upgrade the storage components first (Receive, Store, etc.) a
 - [#8799](https://github.com/thanos-io/thanos/pull/8799): *: Set a `KeepaliveEnforcementPolicy` with `MinTime: 10s` on all gRPC servers, matching the client keepalive interval.
 - [#8806](https://github.com/thanos-io/thanos/pull/8806): Receive: Validate tenant IDs extracted from split-tenant labels to prevent path traversal.
 - [#8810](https://github.com/thanos-io/thanos/pull/8810): Ruler: correctly pass query partial response for gRPC.
+- [#8812](https://github.com/thanos-io/thanos/issues/8812): Receive: Validate tenant IDs from gRPC RemoteWrite requests to prevent path traversal.
 
 ### Added